Moves check_k8s_env to deployment_utils to avoid CI import failure.
